updating navbar conditions
